🗓️ IT-141: Implement reservations to be viewed
